+++ var/httpd/htdocs/js/Core.Agent.Search.js
@@ -67,24 +67,10 @@ Core.Agent.Search = (function (TargetNS) {
                 });
 
             // Register event for tree selection dialog
-            $('.ShowTreeSelection').unbind('click').bind('click', function (Event) {
-                Core.UI.TreeSelection.ShowTreeSelection($(this));
-                return false;
-            });
+            Core.UI.TreeSelection.InitTreeSelection();
 
             // Initially display dynamic fields with TreeMode = 1 correctly
-            $('.DynamicFieldWithTreeView').each(function() {
-                var Data = [];
-                $(this).find('option').each(function() {
-                    Data.push([
-                        $(this).attr('value'),
-                        $(this).text(),
-                        $(this).prop('selected'),
-                        $(this).prop('disabled')
-                    ]);
-                });
-                Core.UI.TreeSelection.RestoreDynamicFieldTreeView($(this), Data, 1);
-            });
+            Core.UI.TreeSelection.InitDynamicFieldTreeViewRestore();
         }
 
         return false;